requirements.txt
torch
torchvision
Pillow
numpy
scikit-learn
scipy
tqdm